+++ Classes/Service/PageService.php
@@ -67,27 +67,6 @@ class PageService
             }
         }
 
-        // veto hook
-        foreach ($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['crawler']['pageVeto'] ?? [] as $key => $func) {
-            trigger_error(
-                'The pageVeto-hook of the TYPO3 Crawler is deprecated since v11.0.0 and will be removed in v13.0,
-                please use the PSR-14 ModifySkipPageEvent instead.',
-                E_USER_DEPRECATED
-            );
-
-            $params = [
-                'pageRow' => $pageRow,
-            ];
-            // expects "false" if page is ok and "true" or a skipMessage if this page should _not_ be crawled
-            $veto = GeneralUtility::callUserFunction($func, $params, $this);
-            if ($veto !== false) {
-                if (is_string($veto)) {
-                    return $veto;
-                }
-                return 'Veto from hook "' . htmlspecialchars((string) $key) . '"';
-            }
-        }
-
         $event = $this->eventDispatcher->dispatch(new ModifySkipPageEvent($pageRow));
         return $event->isSkipped();
     }